+++ src/Driver/LambdatestFactory.php
@@ -81,9 +81,17 @@ final class LambdatestFactory extends FacebookFactory
         if ($browser === WebDriverBrowserType::CHROME &&
             isset($capabilities['chromeOptions']) === true &&
             isset($capabilities['chromeOptions']['w3c']) === true &&
-            $chromeW3c !== null &&
             $capabilities['chromeOptions']['w3c'] !== $chromeW3c) {
+            //Restore the configuration value
             $capabilities['chromeOptions']['w3c'] = $chromeW3c;
+            //If no option in behat configuration, remove option
+            if ($chromeW3c === null) {
+                unset($capabilities['chromeOptions']['w3c']);
+            }
+        }
+
+        if (\array_key_exists('chromeOptions', $capabilities) && \count($capabilities['chromeOptions']) === 0) {
+            unset($capabilities['chromeOptions']);
         }
 
         $def->setArgument(1, $capabilities);